move stack lookup props to -private
